fix: get rofi-pass working in X windows | disable focus-on-mouse-move
EXWM's version of focus on mouse move is annoying; on switching a workspace, the
focus automatically switches to whatever window has the mouse cursor. To make it
even worse, if the mouse happen to be where a company overlay appears, EXWM
decides to consider that to be a separate focused window, and my picom setup
would make the actual emacs window transparent.