fix: keybindings for new changes in meow
It is going to be a annoying to switch to meow's new way of doing things. Looks
like meow is slightly shifting from its earlier approach to modal editing, to
which I had gotten a bit used to. =SPC-g= don't seem possible anymore because
meow has decided =SPC-g= will be converted to =C-M-= now. I believe it can be
configured, but I don't want it to become a time sink of maintenance. So for the
time being, I am gonna try getting used to =SPC-g= being =C-M-=. Let's see how
it works out.