fix: remove leader-key support in exwm
meow is making a mess of it. For example, it converts =SPC b b= to =C-c b C-b=
instead of =C-c b b= which it used to do earlier. I am seeing diminishing
returns for keep investing in meow; it isn't adding enough value to justify
putting up with the nonsense it is creating. I might need to remove it
altogether soon.