feat: add straight.el to support macos
Also add it on Guix to isntall packages which are not present in Guix channels.
I was cloning them manually, but this is a better alternative. Best would be to
create a channel and build those packages with guix itself, but this is fine
until I get around to learn how to package stuff with Guix.